+++ packages/react/src/molecules/dataTable/base/filter.tsx
@@ -1,7 +1,7 @@
 import React, { forwardRef } from "react";
-import TextFilter from "./TextFilter";
-import CheckboxFilter from "./CheckboxFilter";
-import RadioFilter from "./RadioFilter";
+import TextFilter from "./textFilter";
+import CheckboxFilter from "./checkboxFilter";
+import RadioFilter from "./radioFilter";
 
 const Filter = forwardRef(function Filter(
   { item, applyFilter }: { item: any; applyFilter: any },
@@ -12,32 +12,36 @@ const Filter = forwardRef(function Filter(
     <td className="px-6 py-3">
       {filterable && filterable.type === "text" && (
         <TextFilter
+          item={item}
           onChange={(e: any) => {
             applyFilter(e);
           }}
         />
       )}
 
-      {/* {filterable && filterable.type === "checkbox" && (
-        <CheckboxFilter
-          selected={filterable.selected}
-          options={filterable.options}
-          onChange={(e: any) => {
-            applyFilter(e);
-          }}
-        />
+      {filterable && filterable.type === "checkbox" && (
+        <>
+          <CheckboxFilter
+            selected={filterable.selected}
+            options={filterable.options}
+            onChange={(e: any) => {
+              applyFilter(e);
+            }}
+          />
+        </>
       )}
 
       {filterable && filterable.type === "radio" && (
-        <RadioFilter
-          selected={filterable.selected}
-          options={filterable.options}
-          onChange={(e) => {
-            console.log("Radio filter changed", e);
-            applyFilter(e);
-          }}
-        />
-      )} */}
+        <>
+          <RadioFilter
+            selected={filterable.selected}
+            options={filterable.options}
+            onChange={(e: any) => {
+              applyFilter(e);
+            }}
+          />
+        </>
+      )}
 
       {filterable && filterable.type === "clear" && (
         <div className="pr-6 text-right">