+++ packages/react/tailwind.config.js
@@ -5,5 +5,10 @@ module.exports = {
     extend: {},
   },
   variants: {},
-  plugins: [require("@tailwindcss/forms")],
+  plugins: [
+    require("@tailwindcss/forms"),
+    function ({ addVariant }) {
+      addVariant("initial", "html :where(&)");
+    },
+  ],
 };